Add more tests for the TRN lookup
Also moves most of the logic into a TrnLookupHelper class (to simplify
testing).
